requirements.txt
absl_py==2.1.0
Bio==1.7.1
biopandas==0.4.1
biopython==1.81
deepspeed==0.5.9
DLLogger==1.0.0
dm_tree==0.1.7
fair_esm==2.0.0
flash_attn==2.6.3
matplotlib==3.7.2
ml_collections==0.1.1
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
OpenMM==8.1.2
pandas==2.0.3
pdbfixer==1.9.0
pymol==2.5.0
pyrosetta==2023.33+release.9c16e13c3cc
pytorch_lightning==1.8.6
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.14.1
setuptools==72.1.0
torch==2.0.1
torch==2.0.0
transformers==4.43.4
